Final refactor after suggestions thread
