+++ core/src/main/scala/com/raphtory/Raphtory.scala
@@ -5,15 +5,15 @@ import com.raphtory.api.analysis.graphview.TemporalGraph
 import com.raphtory.api.analysis.graphview.TemporalGraphConnection
 import com.raphtory.api.input.GraphBuilder
 import com.raphtory.api.input.Spout
-import com.raphtory.internals.communication.topicRepositories.PulsarAkkaTopicRepository
-import com.raphtory.internals.communication.topicRepositories.PulsarTopicRepository
+import com.raphtory.internals.communication.repositories.PulsarAkkaTopicRepository
+import com.raphtory.internals.communication.repositories.PulsarTopicRepository
 import com.raphtory.internals.components.querymanager.Query
 import com.raphtory.internals.management.ComponentFactory
 import com.raphtory.internals.management.ConfigHandler
+import com.raphtory.internals.management.GraphDeployment
 import com.raphtory.internals.management.MonixScheduler
 import com.raphtory.internals.management.Py4JServer
-import com.raphtory.internals.management.client.GraphDeployment
-import com.raphtory.internals.management.client.QuerySender
+import com.raphtory.internals.management.QuerySender
 import com.raphtory.spouts.IdentitySpout
 import com.typesafe.config.Config
 import com.typesafe.scalalogging.Logger
