+++ core/src/main/scala/com/raphtory/api/analysis/graphview/RaphtoryGraph.scala
@@ -1,7 +1,7 @@
 package com.raphtory.api.analysis.graphview
 
 import com.raphtory.internals.components.querymanager.Query
-import com.raphtory.internals.management.client.QuerySender
+import com.raphtory.internals.management.QuerySender
 
 /** Reduced GraphView with fixed timeline
   *
