+++ core/src/main/scala/com/raphtory/api/analysis/graphview/TemporalGraphConnection.scala
@@ -3,7 +3,7 @@ package com.raphtory.api.analysis.graphview
 import com.raphtory.internals.communication.TopicRepository
 import com.raphtory.internals.components.querymanager.Query
 import com.raphtory.internals.management.MonixScheduler
-import com.raphtory.internals.management.client.QuerySender
+import com.raphtory.internals.management.QuerySender
 import com.typesafe.config.Config
 
 /**  `TemporalGraphConnection` is a wrapper for the `TemporalGraph` class normally used to interact with Raphtory graphs.
