+++ core/src/main/scala/com/raphtory/api/analysis/table/TableImplementation.scala
@@ -3,7 +3,7 @@ package com.raphtory.api.analysis.table
 import com.raphtory.api.output.sink.Sink
 import com.raphtory.internals.components.querymanager.Query
 import com.raphtory.internals.components.querytracker.QueryProgressTracker
-import com.raphtory.internals.management.client.QuerySender
+import com.raphtory.internals.management.QuerySender
 
 /** @note DoNotDocument */
 private[api] class TableImplementation(val query: Query, private val querySender: QuerySender)
